% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{run the Shiny Application}
\usage{
run_app(indexrmd = NA, home = "/", onStart = NULL, uiPattern = "/",
  port = 2811, enableBookmarking = NULL, ...)
}
\arguments{
\item{indexrmd}{the path to index.Rmd, if you want to launch from an already existing bookdown}

\item{home}{where should the directory/file selector widget start from. Default is "/"}

\item{onStart}{A function that will be called before the app is actually run.
This is only needed for \code{shinyAppObj}, since in the \code{shinyAppDir}
case, a \code{global.R} file can be used for this purpose.}

\item{uiPattern}{A regular expression that will be applied to each \code{GET}
request to determine whether the \code{ui} should be used to handle the
request. Note that the entire request path must match the regular
expression in order for the match to be considered successful.}

\item{port}{port to run the Shiny app on, default is 2811.}

\item{enableBookmarking}{Can be one of \code{"url"}, \code{"server"}, or
\code{"disable"}. This is equivalent to calling the
\code{\link{enableBookmarking}()} function just before calling
\code{shinyApp()}. With the default value (\code{NULL}), the app will
respect the setting from any previous calls to \code{enableBookmarking()}.
See \code{\link{enableBookmarking}} for more information.}

\item{...}{options to be passed to `shinyApp`, options param}
}
\description{
run the Shiny Application
}
\examples{

if (interactive()) {

  run_app()

}

}
